+++ core/designsystem/src/main/java/com/uuranus/designsystem/component/TextUtils.kt
@@ -5,7 +5,6 @@ import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.withStyle
 
-
 fun String.toAnnotateString(changeText: String, changedColor: Color): AnnotatedString {
     val text = AnnotatedString.Builder(this)
     text.withStyle(style = SpanStyle(color = changedColor)) {
