+++ feature/home/src/main/java/com/uuranus/home/HomeScreen.kt
@@ -3,24 +3,17 @@ package com.uuranus.home
 import androidx.compose.animation.AnimatedVisibilityScope
 import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.IntrinsicSize
-import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
@@ -37,7 +30,6 @@ import androidx.compose.ui.draw.drawBehind
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -45,19 +37,15 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.uuranus.designsystem.component.CircularImageComponent
 import com.uuranus.designsystem.component.LoadingScreen
 import com.uuranus.designsystem.component.MyScheduleAppBar
-import com.uuranus.designsystem.component.MyScheduleDialog
-import com.uuranus.designsystem.component.MyScheduleFilledButton
 import com.uuranus.designsystem.component.toAnnotateString
 import com.uuranus.designsystem.theme.MyScheduleTheme
 import com.uuranus.home.calendar.DateInfo
 import com.uuranus.home.calendar.ScheduleCalendar
 import com.uuranus.home.calendar.ScheduleData
 import com.uuranus.home.calendar.ScheduleInfo
-import com.uuranus.home.calendar.getLanguageYMDDate
 import com.uuranus.model.MyScheduleInfo
 import com.uuranus.navigation.MyScheduleScreens
 import com.uuranus.navigation.currentComposeNavigator
-import java.util.Date
 
 internal val calendarColors = listOf(
     Color(0xFFF3A8A8),
@@ -78,7 +66,9 @@ fun HomeScreen(
 
     val homeUiState by homeViewModel.homeUiState.collectAsStateWithLifecycle()
 
-    val memberIdColorMap = mutableMapOf<Int, Color>()
+    val memberIdColorMap = remember{
+        mutableMapOf<Int, Color>()
+    }
 
     val currentColorIndex = remember {
         mutableIntStateOf(0)
@@ -262,71 +252,6 @@ fun HomeContent(
 }
 
 @Composable
-fun BottomSheetContent(
-    dateInfo: DateInfo,
-    scheduleInfo: List<ScheduleData<MyScheduleInfo>>,
-    onClick: (DateInfo, ScheduleData<MyScheduleInfo>) -> Unit,
-) {
-
-    LazyColumn(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp)
-    ) {
-        item {
-            Row(
-                modifier = Modifier.fillMaxSize(),
-                horizontalArrangement = Arrangement.Center
-            ) {
-                Text(
-                    text = getLanguageYMDDate(dateInfo),
-                    style = MyScheduleTheme.typography.semiBold16
-                )
-            }
-        }
-        item {
-            Spacer(modifier = Modifier.height(24.dp))
-        }
-        items(scheduleInfo.size) { index ->
-            MyScheduleDetailListItem(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .wrapContentHeight(),
-                scheduleInfo[index]
-            ) {
-                //내가 사장일 때
-
-                //아닐 때
-                if (scheduleInfo[index].detail.isFillInNeeded || scheduleInfo[index].detail.isMine) {
-                    Row {
-                        MyScheduleFilledButton(
-                            modifier = Modifier,
-                            paddingValues = PaddingValues(
-                                horizontal = 14.dp,
-                                vertical = 5.dp
-                            ), buttonState = true,
-                            content = {
-                                if (scheduleInfo[index].detail.isMine) {
-                                    Text(
-                                        "요청", style = MyScheduleTheme.typography.regular14,
-                                    )
-                                } else if (scheduleInfo[index].detail.isFillInNeeded) {
-                                    Text("수락", style = MyScheduleTheme.typography.regular14)
-                                }
-                            }
-                        ) {
-                            onClick(dateInfo, scheduleInfo[index])
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-
-}
-
-@Composable
 fun MyScheduleDetailListItem(
     modifier: Modifier,
     scheduleInfo: ScheduleData<MyScheduleInfo>,
@@ -392,99 +317,6 @@ fun MyScheduleDetailListItem(
 }
 
 
-@Composable
-fun RequestFillInDialog(
-    dateInfo: DateInfo,
-    scheduleInfo: ScheduleData<MyScheduleInfo>,
-    onDismissDialog: () -> Unit,
-) {
-    AlertDialog(
-        onDismissRequest = onDismissDialog,
-        containerColor = MyScheduleTheme.colors.background,
-        title = {
-            Text(
-                "대타 요청",
-                style = MyScheduleTheme.typography.bold20
-            )
-        },
-        text = {
-            Column() {
-                Text(
-                    getLanguageYMDDate(dateInfo = dateInfo),
-                    style = MyScheduleTheme.typography.regular16
-                )
-                Spacer(modifier = Modifier.height(18.dp))
-                MyScheduleDetailListItem(
-                    modifier = Modifier,
-                    scheduleInfo = scheduleInfo
-                )
-            }
-        },
-        confirmButton = {
-            MyScheduleFilledButton(
-                modifier = Modifier
-                    .fillMaxWidth(),
-                paddingValues = PaddingValues(all = 13.dp),
-                buttonState = true,
-                content = {
-                    Text(
-                        "요청",
-                        style = MyScheduleTheme.typography.semiBold16
-                    )
-                }
-            ) {
-                //viewModel.request()
-            }
-        }
-    )
-}
-
-@Composable
-fun AcceptFillInDialog(
-    dateInfo: DateInfo,
-    scheduleInfo: ScheduleData<MyScheduleInfo>,
-    onDismissDialog: () -> Unit,
-) {
-    AlertDialog(
-        onDismissRequest = onDismissDialog,
-        containerColor = MyScheduleTheme.colors.background,
-        title = {
-            Text(
-                "대타 수락",
-                style = MyScheduleTheme.typography.bold20
-            )
-        },
-        text = {
-            Column() {
-                Text(
-                    getLanguageYMDDate(dateInfo = dateInfo),
-                    style = MyScheduleTheme.typography.regular16
-                )
-                Spacer(modifier = Modifier.height(18.dp))
-                MyScheduleDetailListItem(
-                    modifier = Modifier,
-                    scheduleInfo = scheduleInfo
-                )
-            }
-
-        },
-        confirmButton = {
-            MyScheduleFilledButton(
-                modifier = Modifier.fillMaxWidth(),
-                paddingValues = PaddingValues(all = 13.dp),
-                buttonState = true,
-                content = {
-                    Text(
-                        "수락",
-                        style = MyScheduleTheme.typography.semiBold16,
-                    )
-                }
-            ) {
-                //viewModel.accept()
-            }
-        },
-    )
-}
 
 
 @Preview
