+++ feature/home/src/main/java/com/uuranus/home/calendar/ScheduleCalendar.kt
@@ -18,15 +18,15 @@ import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.drawBehind
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.tooling.preview.Preview
@@ -135,7 +135,7 @@ fun <T> ScheduleCalendarMonth(
 
 @Composable
 fun ScheduleCalendarWeekDay(weekNum: Int) {
-    val week = listOf("월", "화", "수", "목", "금", "토", "일")
+    val week = getSystemWeeks()
     val color =
         MyScheduleTheme.colors.gray
