+++ feature/home/src/main/java/com/uuranus/home/calendar/ScheduleCalendarListItem.kt
@@ -5,12 +5,12 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.uuranus.designsystem.theme.MyScheduleTheme
-import androidx.compose.material3.Text
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import com.uuranus.designsystem.theme.MyScheduleTheme
 
 @Composable
 fun <T> ScheduleCalendarListItem(
